refactor(Routes): ADDINg pre-made excpetions to the routes, typing the exceptions to have a better code look
