+++ app/routes/auth.py
@@ -2,17 +2,17 @@ from typing import Annotated
 
 from fastapi import APIRouter
 from fastapi import Depends
-from fastapi import HTTPException
 from fastapi.security import OAuth2PasswordRequestForm
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from app.database import get_session
+from app.core.database import get_session
+from app.core.exceptions import InvalidCredentials
+from app.core.security import create_access_token
+from app.core.security import get_current_user
+from app.core.security import verify_password
 from app.models import User
 from app.schemas.token_schema import Token
-from app.security import create_access_token
-from app.security import get_current_user
-from app.security import verify_password
 
 router = APIRouter(prefix="/auth", tags=["auth"])
 
@@ -29,10 +29,10 @@ def login_for_access_token(
     user = session.scalar(select(User).where(User.email == form_data.username))
 
     if not user:
-        raise HTTPException(status_code=400, detail="Incorrect email or password")
+        raise InvalidCredentials(detail="Incorrect email or password")
 
     if not verify_password(form_data.password, user.password):
-        raise HTTPException(status_code=400, detail="Incorrect email or password")
+        raise InvalidCredentials(detail="Incorrect email or password")
 
     access_token = create_access_token(data={"sub": user.email})
 
@@ -45,4 +45,4 @@ def refresh_access_token(
 ):
     new_access_token = create_access_token(data={"sub": user.email})
 
-    return {"access_token": new_access_token, "token_type": "bearer"}
+    return Token(access_token=new_access_token, token_type="bearer")
