+++ app/routes/users.py
@@ -2,20 +2,19 @@ from typing import Annotated
 
 from fastapi import APIRouter
 from fastapi import Depends
-from fastapi import HTTPException
-from fastapi import status
 from sqlalchemy import select
 from sqlalchemy.orm import Session
 
-from app.database import get_session
+from app.core.database import get_session
+from app.core.exceptions import DuplicatedError
+from app.core.exceptions import InvalidCredentials
+from app.core.security import get_current_user
+from app.core.security import get_password_hash
 from app.models import User
 from app.schemas.base_schemas import Message
 from app.schemas.users_schema import UserList
 from app.schemas.users_schema import UserPublicDto
 from app.schemas.users_schema import UserSchema
-from app.security import get_current_user
-from app.security import get_password_hash
-
 
 router = APIRouter(prefix="/users", tags=["users"])
 
@@ -35,7 +34,7 @@ def create_user(user: UserSchema, session: Sessions):
     db_user = session.scalar(select(User).where(User.username == user.username))
 
     if db_user:
-        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Username already registered")
+        raise DuplicatedError(detail="Username already registered")
 
     hashed_password = get_password_hash(user.password)
     db_user = User(username=user.username, password=hashed_password, email=user.email)
@@ -46,14 +45,15 @@ def create_user(user: UserSchema, session: Sessions):
     return db_user
 
 
+############# CHECCKAR DEPOIS - NO ANTIGO ESTAVA COM EXCEPTION ERRADO NO ID E PASSOU NO TEST
 @router.put("/{id}", response_model=UserPublicDto)
 def update_user(id: int, user: UserSchema, session: Sessions, current_user: CurrentUser):
     if current_user.id != id:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not enough permissions")
+        raise InvalidCredentials(detail="Not enough permissions")
 
     user_exist = session.scalar(select(User).where(User.email == user.email))
     if user_exist:
-        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Username already registered")
+        raise DuplicatedError(detail="Username already registered")
 
     current_user.username = user.username
     current_user.password = get_password_hash(user.password)
@@ -67,9 +67,9 @@ def update_user(id: int, user: UserSchema, session: Sessions, current_user: Curr
 @router.delete("/{id}", response_model=Message)
 def delete_user(id: int, session: Sessions, current_user: CurrentUser):
     if current_user.id != id:
-        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Not enough permissions")
+        raise InvalidCredentials(detail="Not enough permissions")
 
     session.delete(current_user)
     session.commit()
 
-    return {"detail": "User deleted"}
+    return Message(detail="User deleted")
